Be/feature 202/get plant

See merge request s11-webmobile2-sub2/S11P12B308!14

+++ backend/src/main/java/com/plog/backend/domain/plant/entity/OtherPlantType.java
@@ -14,7 +14,7 @@ import lombok.ToString;
 public class OtherPlantType {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int otherPlantTypeId;
+    Long otherPlantTypeId;
     @Column
     String plantName;
 }
